# src/main/java/com/mishchuk/autotrade/controller/dto/CabinetSourceStatusUpdateDto.java
package com.mishchuk.autotrade.controller.dto;

import com.fasterxml.jackson.databind.PropertyNamingStrategies;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import com.mishchuk.autotrade.enums.Status;
import lombok.Data;

@Data
@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
public class CabinetSourceStatusUpdateDto {
    private Status status;
}