# src/main/java/com/mishchuk/autotrade/controller/dto/TransactionCreateDto.java
package com.mishchuk.autotrade.controller.dto;

import com.fasterxml.jackson.databind.PropertyNamingStrategies;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import com.mishchuk.autotrade.enums.Direction;
import lombok.Builder;
import lombok.Data;
import java.math.BigDecimal;
import java.time.Instant;
import java.util.UUID;

@Data
@Builder
@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
public class TransactionCreateDto {
    private final BigDecimal amount;
    private final Direction direction;
    private final Instant createdAt;
    private final UUID accountId;
    private final UUID sourceId;
}