# src/main/java/com/mishchuk/autotrade/service/model/Source.java
package com.mishchuk.autotrade.service.model;

import com.mishchuk.autotrade.enums.Status;
import jakarta.persistence.*;
import lombok.*;
import java.time.Instant;

@Builder
@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Entity
@Table(name = "sources")
public class Source {
    @Id
    @GeneratedValue
    private String id;
    private String name;
    private String platform;
    private String token;
    private Status status;
    protected Instant createdAt;
    protected Instant updatedAt;
}